refactor reactivate user
