+++ frontend/src/app/Settings/Users/ReactivateUser/ReactivateUserModal.tsx
@@ -1,27 +1,26 @@
-import React from "react";
 import Modal from "react-modal";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-import Button from "../../commonComponents/Button/Button";
-import { displayFullName } from "../../utils";
-
-import { SettingsUser } from "./ManageUsersContainer";
-import "./ManageUsers.scss";
+import Button from "../../../commonComponents/Button/Button";
+import { displayFullName } from "../../../utils";
+import { SettingsUser } from "../ManageUsersContainer";
 
 interface Props {
+  isOpen: boolean;
   onClose: () => void;
   onReactivateUser: (userId: string) => void;
   user: SettingsUser;
 }
 
-const ReactivateUserModal: React.FC<Props> = ({
+const ReactivateUserModal = ({
+  isOpen,
   onClose,
   onReactivateUser,
   user,
-}) => {
+}: Props) => {
   return (
     <Modal
-      isOpen={true}
+      isOpen={isOpen}
       style={{
         content: {
           maxHeight: "90vh",
