+++ frontend/src/app/Settings/Users/ReactivateUser/operations.graphql
@@ -1,0 +1,5 @@
+mutation ReactivateUser($id: ID!) {
+  reactivateUser(id: $id) {
+    id
+  }
+}
